+++ service/controller/resource/monitoring/ingress/v1beta1/client.go
@@ -1,4 +1,4 @@
-package ingress
+package v1beta1
 
 import (
 	"context"