+++ service/service.go
@@ -141,6 +141,7 @@ func New(config Config) (*Service, error) {
 				Provider:                provider,
 				Region:                  config.Viper.GetString(config.Flag.Service.Installation.Region),
 				Registry:                config.Viper.GetString(config.Flag.Service.Installation.Registry),
+				IngressApiVersion:       config.Viper.GetString(config.Flag.Service.Installation.IngressApiVersion),
 
 				OpsgenieKey: config.Viper.GetString(config.Flag.Service.Opsgenie.Key),
 
@@ -188,6 +189,7 @@ func New(config Config) (*Service, error) {
 				Provider:                provider,
 				Region:                  config.Viper.GetString(config.Flag.Service.Installation.Region),
 				Registry:                config.Viper.GetString(config.Flag.Service.Installation.Registry),
+				IngressApiVersion:       config.Viper.GetString(config.Flag.Service.Installation.IngressApiVersion),
 
 				OpsgenieKey: config.Viper.GetString(config.Flag.Service.Opsgenie.Key),
 
@@ -233,6 +235,7 @@ func New(config Config) (*Service, error) {
 			Provider:                provider,
 			Region:                  config.Viper.GetString(config.Flag.Service.Installation.Region),
 			Registry:                config.Viper.GetString(config.Flag.Service.Installation.Registry),
+			IngressApiVersion:       config.Viper.GetString(config.Flag.Service.Installation.IngressApiVersion),
 
 			AlertmanagerAddress:     config.Viper.GetString(config.Flag.Service.Alertmanager.Address),
 			AlertmanagerCreatePVC:   config.Viper.GetBool(config.Flag.Service.Alertmanager.Storage.CreatePVC),