+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Support ingress v1 by default.
+
 ## [2.2.0] - 2022-01-20
 
 ### Changed